Create main.cpp with 'Hello World!' requiring C++ 14
Using C++ 14 Lambdas and Net Captures

[out = std::ref(std::cout << "Hello ")](){ out.get() << "World\n"; }();

Initialize the value 'out' as a standard reference wrapper around
the result value of calling the insertion operator on the 'cout'
global variable with the string 'Hello '.

Then our lambda is going to take no parameters and then we need
to dereference the reference wrapper to get back the reference
with the .get() method.

Then we will call 'World' here and then invoke our lambda and
this does what we wanted it to do.